+++ glue/core/tests/test_state.py
@@ -5,10 +5,11 @@ from io import BytesIO
 
 import pytest
 import numpy as np
+from numpy.testing import assert_equal
 
 from glue.external import six
 from glue import core
-from glue.core.component import CategoricalComponent
+from glue.core.component import CategoricalComponent, DateTimeComponent
 from glue.tests.helpers import requires_astropy, make_file
 
 from ..data_factories import load_data
@@ -225,8 +226,8 @@ def test_polygonal_roi():
     roi.add_point(0, 1)
     roi.add_point(1, 0)
     r2 = clone(roi)
-    assert r2.vx == [0, 0, 1]
-    assert r2.vy == [0, 1, 0]
+    assert_equal(r2.vx, [0, 0, 1])
+    assert_equal(r2.vy, [0, 1, 0])
 
 
 def test_matplotlib_cmap():
@@ -241,6 +242,14 @@ def test_categorical_component():
     np.testing.assert_array_equal(c.codes, [0, 1, 2, 0, 1])
     np.testing.assert_array_equal(c.labels, ['a','b','c','a','b'])
     np.testing.assert_array_equal(c.categories, ['a','b','c'])
+
+
+def test_datetime_component():
+    c = DateTimeComponent(np.array([100, 200, 300], dtype='M8[D]'))
+    c2 = clone(c)
+    assert isinstance(c2, DateTimeComponent)
+    np.testing.assert_array_equal(c.data, c2.data)
+    assert isinstance(c2.data[0], np.datetime64)
 
 
 class DummyClass(object):