+++ glue/viewers/histogram/qt/options_widget.py
@@ -28,8 +28,8 @@ class HistogramOptionsWidget(QtWidgets.QWidget):
         viewer_state.add_callback('x_att', self._update_attribute)
 
     def _update_attribute(self, *args):
-        # If at least one of the components is categorical, disable log button
-        log_enabled = not any(comp.categorical for comp in self.viewer_state._get_x_components())
+        # If at least one of the components is categorical or a date, disable log button
+        log_enabled = not any(comp.categorical or comp.datetime for comp in self.viewer_state._get_x_components())
         self.ui.bool_x_log.setEnabled(log_enabled)
         self.ui.bool_x_log_.setEnabled(log_enabled)
         if not log_enabled: